Migrate fjs/types/nominal/proof.f.ts to .f.mjs
`declare const … : unique symbol` has no JavaScript form, so the two
brand carriers move to the sibling types.ts along with the two types
keyed on them, per the migration doc's rule for mixed runtime/type
modules. The proof keeps the expressions themselves -- they are
statement-level demonstrations of what TypeScript does and does not
reject, which cannot be expressed as declarations in types.ts.

The two moved types must be exported for the proof to @import them, so
they are _-prefixed and private by contract; nothing else in the
module's surface changes.

Verified the proof still demonstrates what it claims, by uncommenting
the two comparisons it documents as compile errors: `strA > strB` on a
Nominal and `a < b` on the symbol-intersection brand each still report
TS2469, and the file is otherwise clean.
