+++ fjs/types/nominal/types.ts
@@ -12,3 +12,18 @@
  */
 export type Nominal<N extends string, R extends string, B> =
     symbol & { [k in N]: readonly [R, B] }
+
+// Brand carriers for the comparison experiments in `proof.f.mjs`. They live
+// here because `declare const` / `unique symbol` have no JavaScript form; the
+// proof imports the two types below and keeps the expressions that demonstrate
+// what TypeScript does or does not reject. `_`-prefixed: private by contract.
+
+declare const noCompareBrand: unique symbol
+
+declare const brand: unique symbol
+
+/** A unique-symbol-keyed brand. TypeScript still permits `<` between two of these. */
+export type _SymbolKeyBranded = { [noCompareBrand]: void }
+
+/** A `symbol` intersection brand. `<` on this is TS2469, so the proof only comments it. */
+export type _SymbolIntersectionBranded = symbol & { [brand]: 'SafeId' }
